updated dockerfile with data set

+++ download_dataset.py
@@ -2,7 +2,7 @@ import os
 import tarfile
 import sys
 
-from classifiers.globals import SERIALIZED_DATA_PATH
+from classifiers.globals import SERIALIZED_DATA_PATH, EXTRACTED_DATA_SET_DIR_NAME
 
 TAR_DATA_SET_FILENAME = "cifar-10-python.tar.gz"
 URL = "https://www.cs.toronto.edu/~kriz/" + TAR_DATA_SET_FILENAME
@@ -24,9 +24,13 @@ def extract():
 
 
 if __name__ == "__main__":
-    download_data_set()
-    extract()
-    os.remove(TAR_DATA_SET_FILENAME)
+    if not os.path.exists(EXTRACTED_DATA_SET_DIR_NAME):
+        print ("data set downloading")
+        download_data_set()
+        extract()
+        os.remove(TAR_DATA_SET_FILENAME)
+    else:
+        print("data set loaded")
 
     if not os.path.exists(SERIALIZED_DATA_PATH):
         os.makedirs(SERIALIZED_DATA_PATH)
